requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
feedparser==6.0.11
idna==3.10
pymongo==4.11.3
PyYAML==6.0.2
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
Flask
PyYAML
gunicorn 